Clarifies Arduino logic level behavior
Refines the description of Arduino logic levels, specifically
addressing the behavior in the indeterminate voltage range.

It changes "interpret the signal as either HIGH or LOW, so
there is no way to define the behavior of the Arduino in this
range" to "act unpredictably and can read as ether HIGH or LOW".